Revert "[Core] Add "env_vars" field to runtime_env (#16075)"
This reverts commit 1e245005c9a0e285663945ba8de79b72a4c99fa6.
