+++ python/ray/remote_function.py
@@ -270,16 +270,13 @@ class RemoteFunction:
             accelerator_type)
 
         if runtime_env:
-            runtime_env_dict = runtime_support.RuntimeEnvDict(
-                runtime_env).get_parsed_dict()
+            parsed_runtime_env = runtime_support.RuntimeEnvDict(runtime_env)
+            override_environment_variables = (
+                parsed_runtime_env.to_worker_env_vars(
+                    override_environment_variables))
+            runtime_env_dict = parsed_runtime_env.get_parsed_dict()
         else:
             runtime_env_dict = {}
-
-        if override_environment_variables:
-            logger.warning("override_environment_variables is deprecated and "
-                           "will be removed in Ray 1.5.  Please use "
-                           ".options(runtime_env={'env_vars': {...}}).remote()"
-                           "instead.")
 
         def invocation(args, kwargs):
             if self._is_cross_language:
